requirements.txt
attrs==21.4.0
black==22.1.0
click==8.0.3
cloudpickle==2.0.0
decorator==5.1.1
iniconfig==1.1.1
mypy-extensions==0.4.3
numpy==1.22.3
onnx==1.11.0
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.0
pluggy==1.0.0
protobuf==3.19.4
psutil==5.9.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.0
scipy==1.8.0
six==1.16.0
synr==0.6.0
tomli==2.0.1
tornado==6.1
tvm==0.9.dev783+g0b6125637
typing_extensions==4.1.1
xgboost==1.5.2
